redesign TH rules for simplicity
NeedsCompilation depends on GetModuleGraph, which depends on all the KnownTargets and needs to be rebuilt on every edit. This is O(N) for the number of modules in the workspace, and doesn't scale well in large projects with thousands of modules.

To recover O(1) edits, we give up on tracking which type of linkable to use when loading an interface. Instead, we only build a linkable when explicitly asked so, and only ask when we are loading a module that uses TH. Tradeoffs:

+ O(1) edits
- Every module with a reverse dependency that uses TH is typechecked twice and we store the ModIface twice